pyproject.toml
[project]
name = "mcpbro"
version = "0.1.0"
description = ""
authors = [
    {name = "Jakub Skurek",email = "JakubS@kentico.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "chromadb",
    "sentence-transformers",
    "annotated-types",
    "anyio",
    "blinker",
    "cachelib",
    "cachetools",
    "certifi",
    "click",
    "Flask",
    "Flask-Session",
    "Flask-SocketIO",
    "gevent",
    "google-ai-generativelanguage",
    "google-api-core",
    "google-auth",
    "google-genai",
    "googleapis-common-protos",
    "grpcio",
    "grpcio-status",
    "h11",
    "httpcore",
    "httpx",
    "idna",
    "itsdangerous",
    "Jinja2",
    "MarkupSafe",
    "mcp",
    "msgspec",
    "proto-plus",
    "protobuf",
    "pyasn1",
    "pyasn1_modules",
    "pydantic",
    "pydantic_core",
    "rsa",
    "sniffio",
    "typing_extensions",
    "Werkzeug",
    "chroma-mcp (>=0.2.2,<0.3.0)",
    "uv (>=0.7.2,<0.8.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
packages = [
    { include = "mcpbro", from = "src" },
]

[tool.poetry.group.dev.dependencies]
poethepoet = "^0.34.0"

[tool.poe.tasks]
start = "python -B run.py"

